refactor(cef): carve window lifecycle handlers into window/lifecycle.rs
First sub-PR of the commands/window.rs modularization (Plan 1 in
docs/analysis/ANALYSIS_LARGE_FILE_MODULARIZATION_CANDIDATES_2026_05_28.md).
Pure move — no behavior change.

window.rs → window/mod.rs (git rename), with the close path + HWND
resolution helpers extracted to window/lifecycle.rs:

  close_window, close_window_by_label, resolve_window_hwnd,
  find_main_window, find_all_own_windows, find_own_top_level_window,
  capture_hwnd_for_label, FLOATING_PANE_CLASS_NAME

mod.rs: 1462 → 1133 LOC. The motion / chrome / transparency / meta /
creation handlers stay in mod.rs (later carves).

## Visibility

- close_window / close_window_by_label: `pub`, re-exported unconditionally
  (cross-platform, dispatched by ipc.rs).
- find_own_top_level_window / capture_hwnd_for_label: `pub(crate)`,
  re-exported `#[cfg(windows)]` so browser_pane / client / backend keep
  resolving them as `commands::window::<name>` (those call sites are all
  already `#[cfg(windows)]`).
- resolve_window_hwnd / find_all_own_windows: bumped private → `pub(super)`
  because the motion / transparency handlers still in mod.rs call them;
  imported there via a `#[cfg(windows)] use`.
- find_main_window / FLOATING_PANE_CLASS_NAME: stay private to lifecycle.

## Notes

- Removed the now-unused `use cef::{ImplBrowser, ImplBrowserHost}` from
  mod.rs (only the moved fns used those traits) and added them
  cfg-gated to lifecycle.rs. The moved fns' inline single-trait `use`s
  were partial — `resolve` needs ImplBrowserHost for `window_handle()`,
  `capture` needs ImplBrowser for `host()` — both came from the old
  top-level import; lifecycle.rs now imports both at module scope.
- Dropped a latent unused `GetCurrentProcessId` fn-top import the move
  surfaced in find_all_own_windows (the nested callback has its own).

Verified: `cargo check -p agentmux-cef --release` clean (0 errors; the
remaining warning is pre-existing in browser_pane, unrelated).
