+++ .changesets/1780032038-refactor-cef-extract-window-lifecycle-handlers-into-commands-m2vm.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(cef): extract window lifecycle handlers into commands/window/lifecycle.rs
